feat(update): update user
